requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
crispy-bootstrap4==2023.1
Django==4.2.9
django-bootstrap-v5==1.0.11
django-crispy-forms==2.1
django-filter==23.5
migrate==0.3.8
pillow==10.2.0
psycopg2-binary==2.9.9
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
